Firewall added
Firewall added HTTP server can be accessed from INSIDE and OUTSIDE
OUTSIDE can send HTTPS and HTTP packets only, while INSIDE can ping
Access lists added to multi-layer core switches. However, the devices in the VLANs aren't pingable, so I'm unable to verify the access lists.
Also, I noticed like a loop in the core switches that affects the packet tracer efficiency. There is a PC beside the Firewall. I was testing the connectivity of the firewall with the PC. Everything was working. On the contrary, when I connected the firewall with the multi-layer core switch, Guest2, I'm unable to verify the connectivity due to several possible reasons: loop in core, unpingable devices, and black box testing (I'm not fully aware of the configuration that was made in the core).